choreonoid 1.5 version with many bugs fixed
Former-commit-id: 1424cf6ca597d1e9596359cb10a8d05f36219393

+++ share/project/setup.sh
@@ -1,6 +1,11 @@
 #!/bin/bash
 
-DBNAME=teachingDB.sqlite3
+if [ $# -ne 1 ]; then
+	echo "1 argument (database name) is required." 1>&2
+	exit 1
+fi
+
+DBNAME=$1
 
 rm -i ${DBNAME}
